pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["Reversep <adilet.t.k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.41"
alembic = "^1.16.2"
asyncpg = "^0.30.0"
greenlet = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
